Refactor: server와 front 분리
server: nodejs
front: live-server를 사용한 static server
