+++ front/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/static/style/style.css" />
+    <link rel="stylesheet" href="./style/style.css" />
     <title>FE-kakaopage</title>
     <link
       rel="stylesheet"
@@ -483,6 +483,6 @@
         </article> -->
       </main>
     </div>
-    <script type="module" src="/static/js/init.js"></script>
+    <script type="module" src="js/init.js"></script>
   </body>
 </html>
