+++ front/js/articles/Carousel/contentWrapper.js
