+++ front/js/category/Novel.js
