+++ front/js/constants/message.js
