+++ front/js/constants/variable.js
