+++ front/style/main/carousel.css
